Refactor execution strategies to share code
